FAISS/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pydantic==2.5.0
torch>=1.9.0
transformers>=4.21.0
requests==2.31.0
